+++ changelog.md
@@ -9,6 +9,7 @@
 
 ### Bug fixes
 
+- Fixed forks with same source process name.
 - Fixed `inspect` issue when tasks took more than a day in duration.
 
 ## 1.3.1